Kinda cool idea
I have a cool idea! CTF maps are compressed into an archive type file called a .ctfmap file in which the game extracts whats in the file and uses the mts and conf files respectively. These ctfmap files can be put into a maps folder that is used when you're running a server and the server can randomly pick out a .ctfmap file from that folder. This way it is easier to add and remove maps on your CTF server.