Add Parties Feature (experimental) (#1337)
* Add basic parties

Add parties that mostly work but dont do anything

* Basic parties working

Fix bugs with basic parties

* Add in full draft of parties

Add in draft of party alloc functionality

* fix maxpartysize glitch

Fix glitch where it used MAX_PARTY_SIZE instead of ctf_teams.MAX_PARTY_SIZE

* Party allocation fix

Make it work with any party allocator. Also fix some bugs with invites etc.

* Add better allocator for parties

Use a modified player allocator for the parties

* fix typos

* Add max party size as a setting you can change

* Clean up code

Make code pass luacheck. Fix a small bug. Make some of the unchanging party messages be in a table instead of hardcoded

* Try and fix merge conflicts

* Try and fix merge conflict in features.lua

* Try again to fix merge conflict in features.lua

* More info when player is not allowed to join invite

Notify inviter and not just the invited when the invited was unable to join a party cause it was already too big

* Delete an invite if player tries to accept and party is too big

If a valid invite is sent out but in the mean time the number of players in the sender's party grows, and then the invited tries to accept but is rejected cause the inviter's party is too big, the invite will be deleted.

* Fix difference between files

Add back in a newline at end of file that I deleted, in a random file. So that it is identical to upstream.

* Add newline at end of parties.lua

Make parties.lua fit formatting better with a newline at the end

---------

Co-authored-by: LandarVargan <lonewolf04361@gmail.com>